// packages/ui/types/tables-types.ts
export type enhancedArtists = {
  artistName: string | null;
  id: number;
};

export type LpmData = {
  productId?: string;
  productType?: string;
  productTitle?: string;
  productVersion?: string;
  parentLabel?: string;
  label?: string;
  originalReleaseDate?: Date;
  releaseDate?: Date;
  upc?: string;
  catalog?: string;
  productPriceTier?: string;
  productGenre?: string;
  submissionStatus?: string;
  productCLine?: string;
  productPLine?: string;
  preOrderDate?: Date;
  exclusives?: string;
  explicitLyrics?: string;
  productPlayLink?: string;
  linerNotes?: string;
  primaryMetadataLanguage?: string;
  compilation?: string;
  pdfBooklet?: string;
  timedReleaseDate?: Date;
  timedReleaseMusicServices?: Date;
  lastProcessDate?: Date;
  importDate?: Date;
  createdBy?: string;
  lastModified?: Date;
  submittedAt?: Date;
  submittedBy?: string;
  vevoChannel?: string;
  trackType?: string;
  trackId?: string;
  trackVolume?: boolean;
  trackNumber?: string;
  trackName?: string;
  trackVersion?: string;
  trackDisplayArtist?: string;
  isrc?: string;
  trackPriceTier?: string;
  trackGenre?: string;
  audioLanguage?: string;
  trackCLine?: string;
  trackPLine?: string;
  writersComposers?: string;
  publishersCollectionSocieties?: string;
  withholdMechanicals?: string;
  preOrderType?: string;
  instantGratificationDate?: Date;
  duration?: string;
  sampleStartTime?: string;
  explicitLyricsTrack?: string;
  albumOnly?: string;
  lyrics?: string;
  additionalContributorsPerforming?: string;
  additionalContributorsNonPerforming?: string;
  producers?: string;
  continuousMix?: string;
  continuouslyMixedIndividualSong?: string;
  trackPlayLink?: string;
  productDisplayArtist?: enhancedArtists[];
};